fix: casting/definition bugs

+++ libxtd/array/compare.c
@@ -28,8 +28,8 @@ int int_cmp(const void * const v_1, const void * const v_2)
 
 int long_cmp(const void * const v_1, const void * const v_2)
 {
-    int long_1 = * (long *) v_1;
-    int long_2 = * (long *) v_2;
+    long long_1 = * (long *) v_1;
+    long long_2 = * (long *) v_2;
 
     if (long_1 > long_2)
     {
@@ -47,8 +47,8 @@ int long_cmp(const void * const v_1, const void * const v_2)
 
 int float_cmp(const void * const v_1, const void * const v_2)
 {
-    int float_1 = * (float *) v_1;
-    int float_2 = * (float *) v_2;
+    float float_1 = * (float *) v_1;
+    float float_2 = * (float *) v_2;
 
     if (float_1 > float_2)
     {
@@ -66,8 +66,8 @@ int float_cmp(const void * const v_1, const void * const v_2)
 
 int double_cmp(const void * const v_1, const void * const v_2)
 {
-    int double_1 = * (double *) v_1;
-    int double_2 = * (double *) v_2;
+    double double_1 = * (double *) v_1;
+    double double_2 = * (double *) v_2;
 
     if (double_1 > double_2)
     {
